+++ pkg/analyzer/lib/src/dart/element/type.dart
@@ -139,8 +139,11 @@ class FunctionTypeImpl extends TypeImpl implements FunctionType {
   FunctionTypedElement get element {
     var element = super.element;
     // TODO(scheglov) Can we just construct it with the right element?
-    if (element is FunctionTypeAliasElement) {
-      return element.function;
+    if (element is TypeAliasElement) {
+      var aliasedElement = element.aliasedElement;
+      if (aliasedElement is GenericFunctionTypeElement) {
+        return aliasedElement;
+      }
     }
     return element;
   }