+++ pkg/analyzer/test/src/diagnostics/receiver_of_type_never_test.dart
@@ -422,8 +422,7 @@ void f(Never x) {
 
     assertSimpleIdentifier(
       findNode.simple('foo'),
-      readElement: null,
-      writeElement: null,
+      element: null,
       type: 'Never',
     );
   }
@@ -437,8 +436,7 @@ void f(Never x) {
 
     assertSimpleIdentifier(
       findNode.simple('hashCode'),
-      readElement: objectElement.getGetter('hashCode'),
-      writeElement: null,
+      element: objectElement.getGetter('hashCode'),
       type: 'Never',
     );
   }
@@ -452,14 +450,9 @@ void f(Never x) {
       error(HintCode.DEAD_CODE, 29, 2),
     ]);
 
-    if (hasAssignmentLeftResolution) {
-      assertSimpleIdentifier(
-        findNode.simple('foo'),
-        readElement: null,
-        writeElement: null,
-        type: 'dynamic',
-      );
-    }
+    assertSimpleIdentifierAssignmentTarget(
+      findNode.simple('foo'),
+    );
 
     assertAssignment(
       findNode.assignment('foo += 0'),
@@ -481,8 +474,7 @@ void f(Never x) {
 
     assertSimpleIdentifier(
       findNode.simple('toString'),
-      readElement: objectElement.getMethod('toString'),
-      writeElement: null,
+      element: objectElement.getMethod('toString'),
       type: 'Never',
     );
   }
@@ -496,11 +488,8 @@ void f(Never x) {
       error(HintCode.DEAD_CODE, 28, 2),
     ]);
 
-    assertSimpleIdentifier(
+    assertSimpleIdentifierAssignmentTarget(
       findNode.simple('foo'),
-      readElement: null,
-      writeElement: null,
-      type: 'Never',
     );
 
     assertAssignment(
@@ -525,8 +514,7 @@ void f(Never? x) {
 
     assertSimpleIdentifier(
       findNode.simple('foo'),
-      readElement: null,
-      writeElement: null,
+      element: null,
       type: 'dynamic',
     );
   }
@@ -540,8 +528,7 @@ void f(Never? x) {
 
     assertSimpleIdentifier(
       findNode.simple('hashCode'),
-      readElement: objectElement.getGetter('hashCode'),
-      writeElement: null,
+      element: objectElement.getGetter('hashCode'),
       type: 'int',
     );
   }
@@ -555,8 +542,7 @@ void f(Never? x) {
 
     assertSimpleIdentifier(
       findNode.simple('toString'),
-      readElement: objectElement.getMethod('toString'),
-      writeElement: null,
+      element: objectElement.getMethod('toString'),
       type: 'String Function()',
     );
   }
@@ -623,11 +609,10 @@ void f() {
 
     assertSimpleIdentifier(
       findNode.simple('toString'),
-      readElement: elementMatcher(
+      element: elementMatcher(
         objectElement.getMethod('toString'),
         isLegacy: isNullSafetySdkAndLegacyLibrary,
       ),
-      writeElement: null,
       type: 'String Function()',
     );
   }
@@ -641,11 +626,10 @@ void f() {
 
     assertSimpleIdentifier(
       findNode.simple('hashCode'),
-      readElement: elementMatcher(
+      element: elementMatcher(
         objectElement.getGetter('hashCode'),
         isLegacy: isNullSafetySdkAndLegacyLibrary,
       ),
-      writeElement: null,
       type: 'int',
     );
   }