+++ runtime/vm/object.h
@@ -4336,9 +4336,7 @@ class Field : public Object {
 
   FunctionPtr EnsureInitializerFunction() const;
   FunctionPtr InitializerFunction() const {
-    // We rely on the fact that any loads from the initializer function
-    // are dependent loads and avoid the load-acquire barrier here.
-    return raw_ptr()->initializer_function<std::memory_order_relaxed>();
+    return raw_ptr()->initializer_function<std::memory_order_acquire>();
   }
   void SetInitializerFunction(const Function& initializer) const;
   bool HasInitializerFunction() const;